sdk: Move `Keypair` to `signer` module
